+++ Scr/UI/CanvasWindowUI/CanvasBrush.cs
@@ -2,7 +2,7 @@ using RoboticsTools.Util;
 
 namespace RoboticsTools.UI {
     public interface CanvasBrush {
-        byte value { get; set; }
+        float value { get; set; }
         int size { get; set; }
 
         Vector2Int[] GetPixels(Vector2Int selectedPixel);
